Daniel Shea corrections
On Wed, Nov 13, 2019 at 10:25 PM Shea, Daniel <dbj6@wildcats.unh.edu> wrote:
Chris,

It was great to meet you and Dean today. I am looking forward to our future lunches.

Regarding Dean's email address in Groups.io, it does appear that he was successfully added to the group and with the proper email. Where I had sent the email out this morning, perhaps it had not yet reached him, or ended up in a spam folder. I plan to reach out to the group again after a few days or so in order to confirm that there were no issues receiving Groups.io messages before likely retiring the Yahoo Groups page to ensure that all messages are posted to just one group.

In the interim, I have tracked down the GitHub repository for dtree and have begun reading up on it. I noticed that there were a couple TODOs in which the code would not compile in g++ and <charconv> was not being successfully imported. As an exercise, I decided to resolve the code issues, and the attached ZIP should compile with both g++ and Clang, as well as importing <charconv>

On my Ubuntu system, I had to run the following commands for g++ to work:

add-apt-repository ppa:mhier/libboost-latest
apt update
apt-get --purge remove libboost-all-dev libboost-dev libboost-doc
apt-get install -f
apt autoremove
dpkg --configure -a
apt-get clean
apt-get update
apt-get install libboost1.70-dev
add-apt-repository ppa:ubuntu-toolchain-r/test
apt-get update
apt-get install g++-9

The Makefile uses g++-9 instead of g++ (which is actually g++-7 by default on my Ubuntu build). If it helps making development easier, you could also consider adding a Docker file to the repository, which basically sets up all dependencies such that, whenever anyone runs the code in the Docker image, their dependencies and setup are synchronized with everyone else's. This may prevent issues with having to make sure paths and library versions and other such issues don't produce different results on different systems or when different people run it.

Regarding the problem in which the inverse of Maybe in interval form is not equal to itself, one of the proposed workarounds is to use the complement rather than the NOT operator. With my bias in mind of avoiding undefined behavior whenever possible, fixing the complement of (0<=x::x<=1) to be equal to itself, if I'm understanding correctly, is an approach not entirely without precedent; SQL itself operates on three-valued logic, in which NULL is their keyword for the Unknown/Maybe case. When a SQL WHERE clause is written "<column> = NULL", the query will always return zero rows. To work around the case where rows are populated with NULL values and those values should be retrieved, SQL applies the IS operator "<column> IS NULL". Though semantically confusing for any SQL newcomer, it is a convention to which they must adapt in order to use the language effectively. Assuming this, the challenge then becomes more of a technical one in terms of how to distinguish ! versus ~ for an arbitrary variable 'a' when 'a' had previously been assigned to any of the three possible values, or whether to use ~ for all cases and force a SQL IS NULL-esque convention. Anyway, just an initial thought from my first reading. I had a number of other questions but managed to find an answer either later in the documentation or by looking through the code.

In any case, I will next look into downloading a full copy of the Mega site and will let you know if I run into any issues or when I have sent the request for domain control. I hope to do so sometime this week.

Best,
Dan